+++ composeApp/src/androidMain/kotlin/com/dnfapps/arrmatey/ui/sheets/EditMediaSheet.kt
@@ -1,0 +1,80 @@
+package com.dnfapps.arrmatey.ui.sheets
+
+import androidx.compose.runtime.Composable
+import com.dnfapps.arrmatey.arr.api.model.ArrMedia
+import com.dnfapps.arrmatey.arr.api.model.ArrMovie
+import com.dnfapps.arrmatey.arr.api.model.ArrSeries
+import com.dnfapps.arrmatey.arr.api.model.Arrtist
+import com.dnfapps.arrmatey.arr.api.model.Audiobook
+import com.dnfapps.arrmatey.arr.api.model.Author
+import com.dnfapps.arrmatey.arr.api.model.MockMedia
+import com.dnfapps.arrmatey.arr.api.model.QualityProfile
+import com.dnfapps.arrmatey.arr.api.model.RootFolder
+import com.dnfapps.arrmatey.arr.api.model.SearchAudiobook
+import com.dnfapps.arrmatey.arr.api.model.Tag
+
+@Composable
+fun EditMediaSheet(
+    item: ArrMedia,
+    qualityProfiles: List<QualityProfile>,
+    rootFolders: List<RootFolder>,
+    tags: List<Tag>,
+    editInProgress: Boolean,
+    onEditItem: (ArrMedia) -> Unit,
+    onDismiss: () -> Unit
+) {
+    when (item) {
+        is ArrMovie -> EditMovieSheet(
+            item = item,
+            qualityProfiles = qualityProfiles,
+            rootFolders = rootFolders,
+            tags = tags,
+            editInProgress = editInProgress,
+            onEditItem = onEditItem,
+            onDismiss = onDismiss,
+        )
+
+        is ArrSeries -> EditSeriesSheet(
+            item = item,
+            qualityProfiles = qualityProfiles,
+            rootFolders = rootFolders,
+            tags = tags,
+            editInProgress = editInProgress,
+            onEditItem = onEditItem,
+            onDismiss = onDismiss
+        )
+
+        is Arrtist -> EditArtistSheet(
+            item = item,
+            qualityProfiles = qualityProfiles,
+            rootFolders = rootFolders,
+            tags = tags,
+            editInProgress = editInProgress,
+            onEditItem = onEditItem,
+            onDismiss = onDismiss
+        )
+
+        is Author -> EditAuthorSheet(
+            item = item,
+            qualityProfiles = qualityProfiles,
+            rootFolders = rootFolders,
+            tags = tags,
+            editInProgress = editInProgress,
+            onEditItem = onEditItem,
+            onDismiss = onDismiss
+        )
+
+        is Audiobook -> EditAudiobookSheet(
+            item = item,
+            qualityProfiles = qualityProfiles,
+            rootFolders = rootFolders,
+            editInProgress = editInProgress,
+            onEditItem = onEditItem,
+            onDismiss = onDismiss
+        )
+
+        is SearchAudiobook,
+        is MockMedia -> {
+        }
+    }
+}
